+++ temba/api/v2/tests/test_boundaries.py
@@ -1,7 +1,6 @@
 from django.urls import reverse
 
 from temba.locations.models import BoundaryAlias
-from temba.tests import matchers
 
 from . import APITest
 
@@ -23,7 +22,6 @@ class BoundariesEndpointTest(APITest):
         self.state1.geometry = {"type": "MultiPolygon", "coordinates": [[[[1, 1], [1, -1], [-1, -1], [-1, 1], [1, 1]]]]}
         self.state1.save()
 
-        # test without geometry
         self.assertGet(
             endpoint_url,
             [self.user, self.editor, self.admin],
@@ -34,7 +32,10 @@ class BoundariesEndpointTest(APITest):
                     "parent": {"osm_id": "171496", "name": "Rwanda"},
                     "level": 1,
                     "aliases": ["Kigali", "Kigari"],
-                    "geometry": None,
+                    "geometry": {
+                        "type": "MultiPolygon",
+                        "coordinates": [[[[1.0, 1.0], [1.0, -1.0], [-1.0, -1.0], [-1.0, 1.0], [1.0, 1.0]]]],
+                    },
                 },
                 {
                     "osm_id": "171113181",
@@ -101,35 +102,6 @@ class BoundariesEndpointTest(APITest):
                     "aliases": [],
                     "geometry": None,
                 },
-            ],
-            num_queries=self.BASE_SESSION_QUERIES + 3,
-        )
-
-        # test with geometry
-        self.assertGet(
-            endpoint_url + "?geometry=true",
-            [self.admin],
-            results=[
-                {
-                    "osm_id": "1708283",
-                    "name": "Kigali City",
-                    "parent": {"osm_id": "171496", "name": "Rwanda"},
-                    "level": 1,
-                    "aliases": ["Kigali", "Kigari"],
-                    "geometry": {
-                        "type": "MultiPolygon",
-                        "coordinates": [[[[1.0, 1.0], [1.0, -1.0], [-1.0, -1.0], [-1.0, 1.0], [1.0, 1.0]]]],
-                    },
-                },
-                matchers.Dict(),
-                matchers.Dict(),
-                matchers.Dict(),
-                matchers.Dict(),
-                matchers.Dict(),
-                matchers.Dict(),
-                matchers.Dict(),
-                matchers.Dict(),
-                matchers.Dict(),
             ],
             num_queries=self.BASE_SESSION_QUERIES + 3,
         )